# api/openapi.yaml
openapi: 3.1.0
info:
  title: CellBridge Gateway API
  version: 1.0.0
  description: API v1 for a user-owned cellular gateway. Mutating requests require a device token and Idempotency-Key.
servers:
  - url: https://{gateway}/api/v1
    variables:
      gateway:
        default: gateway.example.com
security:
  - bearerAuth: []
paths:
...
  /auth/refresh:
    post:
      security: []
      operationId: refresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          description: Rotated device credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeviceCredentials' }
...
  /messages:
    get:
      operationId: messages
      parameters:
        - { $ref: '#/components/parameters/After' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Message' }
    post:
      operationId: sendMessage
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SendMessage' }
      responses:
        '201':
          description: Queued message
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Message' }
...
  /calls:
    get:
      operationId: listCalls
      parameters: [ { $ref: '#/components/parameters/Limit' } ]
      responses:
        '200':
          description: Recent calls
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Call' }
    post:
      operationId: dial
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DialRequest' }
      responses:
        '201':
          description: Created call
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Call' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: device-access-token }
  parameters:
    After: { name: after, in: query, schema: { type: integer, minimum: 0, default: 0 } }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500, default: 100 } }
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, format: uuid } }
...
  schemas:
...
    DeviceCredentials:
      type: object
      required: [deviceId, accessToken, refreshToken]
      properties:
        deviceId: { type: string }
        accessToken: { type: string }
        refreshToken: { type: string }
    Message:
      type: object
      required: [id, threadKey, direction, peer, body, status, createdAt]
      properties:
        id: { type: string }
        threadKey: { type: string }
        direction: { type: string, enum: [inbound, outbound] }
        peer: { type: string }
        body: { type: string }
        encoding: { type: string, enum: [gsm7, ucs2] }
        status: { type: string, enum: [queued, submitted, sent, delivered, failed, read] }
        createdAt: { type: integer }
...
    SendMessage:
      type: object
      required: [to, body]
      properties:
        to: { type: string, minLength: 1, maxLength: 32 }
        body: { type: string, minLength: 1, maxLength: 10000 }
    DialRequest:
      type: object
      required: [to, clientCallId]
      properties:
        to: { type: string, minLength: 1, maxLength: 32 }
        clientCallId: { type: string, format: uuid }
    Call:
      type: object
      required: [id, direction, state, startedAt]
      properties:
        id: { type: string }
        direction: { type: string, enum: [inbound, outbound] }
        peer: { type: string }
        state: { type: string, enum: [idle, incoming_ringing, outgoing_dialing, connecting, active, ending, recovering] }
        startedAt: { type: integer }
        connectedAt: { type: [integer, 'null'] }
        endedAt: { type: [integer, 'null'] }
        endReason: { type: string }
        recordingId: { type: string }
        recordingState: { type: string }
        recordingDurationMs: { type: integer }
...